requirements.txt
wandb
torch
torchvision
numpy
pandas
scikit-learn
scikit-image
pillow
matplotlib
python-box